Move helpers to service folder
